projects/palantir-foundry-play/requirements.txt
foundry-platform-sdk==0.116.0
python-dotenv==1.1.0
